+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Added`SpcConc%Units` and `SpcConc%Previous_Units` for species-specific unit conversion
 - Diel and day-of-week scale factors for CEDS global base emissions
+`State_Chm%all
+- SpeciesInDryMixingRatio` logical flag, needed to check if all species are in dry mixing ratio before calling AIRQNT with `update_mixing_ratio=.TRUE.`
 
 ### Changed
 - Updated routines in `GeosUtil/unitconv_mod.F90` for species-specific unit conversion